Split ai-agent-runner into focused modules and remove vector/embedder services
Break the 1094-line ai-agent-runner.ts into 5 focused files:
- ai-agent-runner.ts (runAIAgent only)
- ai-agent-stream.ts (streamAIAgent + persisting channel)
- ai-agent-prepare.ts (prepareAgentRun, buildToolDefs, types)
- ai-agent-memory.ts (memory helpers, message persistence)
- ai-agent-registry.ts (addAIAgent, approveAIAgent, getters)

Remove AIVectorService, AIEmbedderService, and semanticRecall from
the agent memory system — too much complexity for the agent MVP.
