--- packages/core/src/services/ai-embedder-service.ts
@@ -1,3 +1,0 @@
-export interface AIEmbedderService {
-  embed(texts: string[]): Promise<number[][]>
-}
