+++ packages/core/src/wirings/ai-agent/index.ts
@@ -1,2 +1,14 @@
 export * from './ai-agent.types.js'
-export * from './ai-agent-runner.js'
+export { runAIAgent } from './ai-agent-runner.js'
+export { streamAIAgent } from './ai-agent-stream.js'
+export {
+  RunAIAgentParams,
+  StreamAIAgentOptions,
+  ToolApprovalRequired,
+} from './ai-agent-prepare.js'
+export {
+  addAIAgent,
+  approveAIAgent,
+  getAIAgents,
+  getAIAgentsMeta,
+} from './ai-agent-registry.js'
